Add static analysis tools Spotbugs and Checkstyle and update gradle build script and fix bugs

+++ src/test/java/com/crowdin/cli/commands/functionality/ProjectUtilsTest.java
@@ -15,7 +15,14 @@ import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.any;
+import static org.mockito.Mockito.eq;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoMoreInteractions;
+import static org.mockito.Mockito.verifyZeroInteractions;
+import static org.mockito.Mockito.when;
 
 public class ProjectUtilsTest {
 
@@ -25,9 +32,9 @@ public class ProjectUtilsTest {
     public void testCreatePath_PathExists() {
         Client client = mock(Client.class);
         Map<String, Long> directoriesIdMap = new HashMap<String, Long>() {{
-            put(Utils.normalizePath("folder/"), 101L);
-            put(Utils.normalizePath("folder/folder2/"), 102L);
-        }};
+                put(Utils.normalizePath("folder/"), 101L);
+                put(Utils.normalizePath("folder/folder2/"), 102L);
+            }};
         String filePath = Utils.normalizePath("folder/folder2/file.txt");
         Branch branch = null;
 
@@ -44,17 +51,17 @@ public class ProjectUtilsTest {
         String filePath = Utils.normalizePath("folder/folder2/file.txt");
         Branch branch = null;
         AddDirectoryRequest request1 = new AddDirectoryRequest() {{
-            setName("folder");
-        }};
+                setName("folder");
+            }};
         AddDirectoryRequest request2 = new AddDirectoryRequest() {{
-            setName("folder2");
-            setDirectoryId(101L);
-        }};
+                setName("folder2");
+                setDirectoryId(101L);
+            }};
         when(client.addDirectory(eq(request1)))
-            .thenReturn(DirectoryBuilder.standard().setProjectId((long) PROJECT_ID)
+            .thenReturn(DirectoryBuilder.standard().setProjectId(PROJECT_ID)
                 .setIdentifiers("folder", 101L, null, null).build());
         when(client.addDirectory(eq(request2)))
-            .thenReturn(DirectoryBuilder.standard().setProjectId((long) PROJECT_ID)
+            .thenReturn(DirectoryBuilder.standard().setProjectId(PROJECT_ID)
                 .setIdentifiers("folder2", 102L, 101L, null).build());
 
         long resultDirectoryId = ProjectUtils.createPath(client, directoriesIdMap, filePath, branch, false);
@@ -69,16 +76,16 @@ public class ProjectUtilsTest {
     public void testCreatePath_HalfPathNotExists() throws ResponseException {
         Client client = mock(Client.class);
         Map<String, Long> directoriesIdMap = new HashMap<String, Long>() {{
-            put(Utils.normalizePath("folder/"), 101L);
-        }};
+                put(Utils.normalizePath("folder/"), 101L);
+            }};
         String filePath = Utils.normalizePath("folder/folder2/file.txt");
         Branch branch = null;
         AddDirectoryRequest request = new AddDirectoryRequest() {{
-            setName("folder2");
-            setDirectoryId(101L);
-        }};
+                setName("folder2");
+                setDirectoryId(101L);
+            }};
         when(client.addDirectory(eq(request)))
-            .thenReturn(DirectoryBuilder.standard().setProjectId((long) PROJECT_ID)
+            .thenReturn(DirectoryBuilder.standard().setProjectId(PROJECT_ID)
                 .setIdentifiers("folder2", 102L, 101L, null).build());
 
         long resultDirectoryId = ProjectUtils.createPath(client, directoriesIdMap, filePath, branch, false);
@@ -92,9 +99,9 @@ public class ProjectUtilsTest {
     public void testCreatePath_PathExists_WithBranch() {
         Client client = mock(Client.class);
         Map<String, Long> directoriesIdMap = new HashMap<String, Long>() {{
-            put(Utils.normalizePath("branch/folder/"), 101L);
-            put(Utils.normalizePath("branch/folder/folder2/"), 102L);
-        }};
+                put(Utils.normalizePath("branch/folder/"), 101L);
+                put(Utils.normalizePath("branch/folder/folder2/"), 102L);
+            }};
         String filePath = Utils.normalizePath("folder/folder2/file.txt");
         Branch branch = BranchBuilder.standard().setProjectId(PROJECT_ID).setIdentifiers("branch", 301L).build();
 
@@ -105,23 +112,24 @@ public class ProjectUtilsTest {
     }
 
     @Test
-    public void testCreatePath_PathNotExists_WithBranch() throws ResponseException{
+    public void testCreatePath_PathNotExists_WithBranch() throws ResponseException {
         Client client = mock(Client.class);
         Map<String, Long> directoriesIdMap = new HashMap<String, Long>();
         String filePath = Utils.normalizePath("folder/folder2/file.txt");
         Branch branch = BranchBuilder.standard().setProjectId(PROJECT_ID).setIdentifiers("branch", 301L).build();
         AddDirectoryRequest request1 = new AddDirectoryRequest() {{
-            setName("folder");
-            setBranchId(branch.getId());
-        }};
+                setName("folder");
+                setBranchId(branch.getId());
+            }};
         AddDirectoryRequest request2 = new AddDirectoryRequest() {{
-            setName("folder2");
-            setDirectoryId(101L);}};
+                setName("folder2");
+                setDirectoryId(101L);
+            }};
         when(client.addDirectory(eq(request1)))
-            .thenReturn(DirectoryBuilder.standard().setProjectId((long) PROJECT_ID)
+            .thenReturn(DirectoryBuilder.standard().setProjectId(PROJECT_ID)
                 .setIdentifiers("folder", 101L, null, branch.getId()).build());
         when(client.addDirectory(eq(request2)))
-            .thenReturn(DirectoryBuilder.standard().setProjectId((long) PROJECT_ID)
+            .thenReturn(DirectoryBuilder.standard().setProjectId(PROJECT_ID)
                 .setIdentifiers("folder2", 102L, 101L, branch.getId()).build());
 
         long resultDirectoryId = ProjectUtils.createPath(client, directoriesIdMap, filePath, branch, false);
@@ -133,18 +141,19 @@ public class ProjectUtilsTest {
     }
 
     @Test
-    public void testCreatePath_HalfPathNotExists_WithBranch() throws ResponseException{
+    public void testCreatePath_HalfPathNotExists_WithBranch() throws ResponseException {
         Client client = mock(Client.class);
         Map<String, Long> directoriesIdMap = new HashMap<String, Long>() {{
-            put(Utils.normalizePath("branch/folder/"), 101L);
-        }};
+                put(Utils.normalizePath("branch/folder/"), 101L);
+            }};
         String filePath = Utils.normalizePath("folder/folder2/file.txt");
         Branch branch = BranchBuilder.standard().setProjectId(PROJECT_ID).setIdentifiers("branch", 301L).build();
         AddDirectoryRequest request = new AddDirectoryRequest() {{
-            setName("folder2");
-            setDirectoryId(101L);}};
+                setName("folder2");
+                setDirectoryId(101L);
+            }};
         when(client.addDirectory(eq(request)))
-            .thenReturn(DirectoryBuilder.standard().setProjectId((long) PROJECT_ID)
+            .thenReturn(DirectoryBuilder.standard().setProjectId(PROJECT_ID)
                 .setIdentifiers("folder2", 102L, 101L, branch.getId()).build());
 
         long resultDirectoryId = ProjectUtils.createPath(client, directoriesIdMap, filePath, branch, false);
@@ -155,7 +164,7 @@ public class ProjectUtilsTest {
     }
 
     @Test
-    public void testCreatePath_PathNotExists_ResponseException() throws ResponseException{
+    public void testCreatePath_PathNotExists_ResponseException() throws ResponseException {
         Client client = mock(Client.class);
         Map<String, Long> directoriesIdMap = new HashMap<String, Long>();
         String filePath = Utils.normalizePath("folder/folder2/file.txt");
@@ -175,11 +184,11 @@ public class ProjectUtilsTest {
         String filePath = Utils.normalizePath("folder/file.txt");
         Branch branch = null;
         AddDirectoryRequest request1 = new AddDirectoryRequest() {{
-            setName("folder");
-        }};
+                setName("folder");
+            }};
         when(client.addDirectory(eq(request1)))
             .thenThrow(new WaitResponseException())
-            .thenReturn(DirectoryBuilder.standard().setProjectId((long) PROJECT_ID)
+            .thenReturn(DirectoryBuilder.standard().setProjectId(PROJECT_ID)
                 .setIdentifiers("folder", 101L, null, null).build());
 
         long resultDirectoryId = ProjectUtils.createPath(client, directoriesIdMap, filePath, branch, false);
